start unit tests

+++ src/java/com/allan/amca/data/DaoAbstract.java
@@ -6,21 +6,33 @@ public abstract class DaoAbstract<T, N> implements Dao<T, N> {
 
     @Override
     public final boolean update(T client, N id) throws SQLException {
+        if (client == null || id == null) {
+            throw new IllegalArgumentException("Neither arguments can be null");
+        }
         return executeUpdate(client, id);
     }
 
     @Override
-    public final boolean delete(T toDelete) throws SQLException {
+    public final boolean delete( T toDelete) {
+        if (toDelete == null) {
+            throw new IllegalArgumentException("Argument cannot be null");
+        }
         return deleteRecord(toDelete);
     }
 
     @Override
     public final boolean create(T toCreate) throws SQLException {
+        if (toCreate == null) {
+            throw new IllegalArgumentException("Argument cannot be null");
+        }
         return addRecord(toCreate);
     }
 
     @Override
     public final <T> T retrieve(N toRetrieve) {
+        if (toRetrieve == null) {
+            throw new IllegalArgumentException("Argument cannot be null");
+        }
         return readRecord(toRetrieve);
     }
